refactor: Applying the Factory design pattern to create new app connections
